Timestamp checks in make.sh

+++ make.sh
@@ -3,9 +3,18 @@
 mklatex () {
     dir=${1%/*}
     tex=${1##*/}
+    out=${tex%.*}.png
     pushd "$dir"
-    pdflatex -shell-escape "$tex"
-    rm -f *.{aux,log,dvi,pdf,ps}
+    update=0
+    for file in *.*
+    do
+        if test "$file" != "$out" -a "$file" -nt "$out"
+        then
+            pdflatex -shell-escape "$tex"
+            rm -f *.{aux,log,dvi,pdf,ps}
+            break
+        fi
+    done
     popd
 }
